requirements.txt
python-dotenv
langchain-nvidia-ai-endpoints==0.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.0.352
unstructured[all-docs]==0.11.2
sentence-transformers==2.2.2
llama-index
dataclass-wizard==0.22.2
opencv-python==4.8.0.74
llama-hub==0.0.43
pymilvus==2.3.1
jupyterlab==4.0.8
langchain-nvidia-trt==0.0.1rc0
langchain-core==0.1.3
atlassian-python-api==3.41.4
requests
llama-index-embeddings-nvidia
llama-index-llms-nvidia      
gradio
pypdf2
docx2txt
llama_index[langchain]
llama-index
